BP16010
tabla Medicamentos

+++ src/main/java/com/proyectoDos/proyecto/models/Usuario.java
@@ -9,14 +9,13 @@ import javax.persistence.*;
 
 @Entity
 @Table(name = "usuarios")
-@ToString
-@EqualsAndHashCode
+@ToString @EqualsAndHashCode
 public class Usuario {
 
     @Id
     @GeneratedValue(strategy=GenerationType.IDENTITY)
     @Getter @Setter @Column(name = "id")
-    private Integer id;
+    private Long id;
 
     @Getter @Setter @Column(name = "nombre")
     private String nombre;
@@ -32,4 +31,6 @@ public class Usuario {
 
     @Getter @Setter @Column(name = "password")
     private String password;
+
+
 }
